requirements.txt
instagrapi==1.17.6
pillow
newsapi-python
bs4
python-dotenv
requests